move slugified
