+++ lib/compile.js
@@ -1,7 +1,6 @@
 import cds from "@sap/cds"
-import { getDescription, getFilteredEntities, getFilteredActions } from "./utils/utils.js"
+import { getDescription, getFilteredEntities, getFilteredActions, slugified } from "./utils/utils.js"
 import { buildAgentCard } from "./protocol/agent-card.js"
-import { slugified } from "./utils/markdown.js"
 
 const A2A_BASE_PATH = "/a2a"
