+++ lib/protocol/agent-card.js
@@ -1,13 +1,12 @@
 import cds from "@sap/cds"
 import { createRequire } from "node:module"
 const { path } = cds.utils
-import { getDescription, getFilteredEntities, getFilteredActions } from "../utils/utils.js"
+import { getDescription, getFilteredEntities, getFilteredActions, slugified } from "../utils/utils.js"
 import {
   scanSkills,
   parseAgentMetadata,
   parseAgentCardMd,
   parseAgentCardFile,
-  slugified,
 } from "../utils/markdown.js"
 
 const require = createRequire(import.meta.url)
