+++ lib/sidecar.js
@@ -1,6 +1,6 @@
 import cds from "@sap/cds"
 import A2AProtocolAdapter, { authContext } from "./index.js"
-import { slugified } from "./utils/markdown.js"
+import { slugified } from "./utils/utils.js"
 import registerDefaultAgentHandlers from "../srv/handlers/index.js"
 
 const LOG = cds.log("a2a")
