+++ lib/utils/markdown.js
@@ -1,17 +1,7 @@
 import cds from "@sap/cds"
+import { slugified } from './utils.js'
 const { path, fs } = cds.utils
 const LOG = cds.log("agents")
-
-/**
- * Mirrors CAP's internal slug rules used for service path generation.
- */
-export const slugified = (name) =>
-  /[^.]+$/
-    .exec(name)[0]
-    .replace(/Service$/, "")
-    .replace(/_/g, "-")
-    .replace(/([a-z0-9])([A-Z])/g, (_m, c, C) => c + "-" + C)
-    .toLowerCase()
 
 /**
  * Absolute filesystem directory of the `.cds` source file that defines `srv`.
