+++ lib/utils/utils.js
@@ -17,6 +17,18 @@ export function getAgentLogger(srv) {
   return cds.log("agents:" + srv.name, { label: agentSlug })
 }
 
+/**
+ * Mirrors CAP's internal slug rules used for service path generation.
+ */
+export const slugified = (name) =>
+  /[^.]+$/
+    .exec(name)[0]
+    .replace(/Service$/, "")
+    .replace(/_/g, "-")
+    .replace(/([a-z0-9])([A-Z])/g, (_m, c, C) => c + "-" + C)
+    .toLowerCase()
+
+
 export function getDescription(obj, locale) {
   locale = locale || cds.context?.locale || "en"
